fabric_tools/pyproject.toml
[project]
name = "fabric-tools"
version = "0.1.0"
description = "Standalone Fabric Graph + Telemetry query tools (extracted from PathfinderIQ)"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "azure-identity>=1.19.0",
    "azure-kusto-data>=4.6.0",
    "agent-framework>=1.0.0rc1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
prerelease = "allow"
